Organize cert-d-java classes in packages
