+++ pgp-cert-d-java-jdbc-sqlite-lookup/src/main/java/pgp/cert_d/jdbc/sqlite/DatabaseSubkeyLookup.java
@@ -11,7 +11,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import pgp.cert_d.SubkeyLookup;
+import pgp.cert_d.subkey_lookup.SubkeyLookup;
 
 public class DatabaseSubkeyLookup implements SubkeyLookup {
