+++ pgp-cert-d-java/src/main/java/pgp/cert_d/backend/InMemoryCertificateDirectoryBackend.java
@@ -2,8 +2,12 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package pgp.cert_d;
+package pgp.cert_d.backend;
 
+import pgp.cert_d.PGPCertificateDirectory;
+import pgp.cert_d.SpecialNames;
+import pgp.cert_d.exception.BadDataException;
+import pgp.cert_d.exception.BadNameException;
 import pgp.certificate.Certificate;
 import pgp.certificate.KeyMaterial;
 import pgp.certificate.KeyMaterialMerger;
