+++ pgp-cert-d-java/src/main/java/pgp/cert_d/backend/package-info.java
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2022 Paul Schaub <vanitasvitae@fsfe.org>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Storage Backends.
+ */
+package pgp.cert_d.backend;
