+++ pgp-cert-d-java/src/main/java/pgp/cert_d/exception/BadDataException.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package pgp.cert_d;
+package pgp.cert_d.exception;
 
 /**
  * The data was not a valid OpenPGP cert or key in binary format.
