+++ pgp-cert-d-java/src/main/java/pgp/cert_d/exception/BadNameException.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package pgp.cert_d;
+package pgp.cert_d.exception;
 
 /**
  * Provided name was neither a valid fingerprint, nor a known special name.
