+++ pgp-cert-d-java/src/main/java/pgp/cert_d/subkey_lookup/package-info.java
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2022 Paul Schaub <vanitasvitae@fsfe.org>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Subkey Lookup functionality.
+ */
+package pgp.cert_d.subkey_lookup;
